my_django/pyproject.toml
[project]
name = "my-django"
version = "0.1.0"
description = ""
authors = [
    {name = "sherbet912",email = "chouchuanyi12@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-admin-interface (>=0.30.1,<0.31.0)",
    "django-ckeditor (>=6.7.3,<7.0.0)",
    "django-markdownx (>=4.0.9,<5.0.0)",
    "django-mdeditor (>=0.1.20,<0.2.0)",
    "markdown (>=3.8.2,<4.0.0)"
]

[tool.poetry]
packages = [{include = "my_django", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
